requirements.txt
flask
flask-restful
pymongo
grequests
python-dotenv